% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acronym.R
\name{acronym}
\alias{acronym}
\title{Create lowercase acronym}
\usage{
acronym(x)
}
\arguments{
\item{x}{A character vector with one element}
}
\value{
A character vector
}
\description{
Create lowercase acronym
}
\examples{
x <- "Robert Hertz"
acronym(x)
}
