\name{AtelieR-internal}
\title{Internal AtelieR objects}
\alias{.First.lib}
\alias{.ws}
\alias{.ws1}
\alias{.ws2}
\alias{.ws3}
\alias{.ws4}
\alias{.ws5}
\alias{.ws6}
\alias{.ws8}
\alias{.ws9}
\alias{.ws10}
\alias{.ws11}
\alias{.ws13}
\description{Internal AtelieR objects.}
\details{These are not to be called by the user.}
\keyword{internal}
