% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atelier-package.R
\docType{package}
\name{atelier-package}
\alias{atelier}
\alias{atelier-package}
\title{atelier: Various Functions For Various Projects}
\description{
Provides personnal shortcuts for data wrangling, plotting, and statitics, based on the tidyverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tblk/atelier}
  \item Report bugs at \url{https://github.com/tblk/atelier/issues}
}

}
\author{
\strong{Maintainer}: Théo Boulakia \email{first.last@example.com}

}
\keyword{internal}
