\name{AtelieR}
\alias{AtelieR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to construct and launch the AtelieR GUI.
}
\description{
A function to construct and launch the AtelieR GUI.
}
\usage{
AtelieR()
}
\details{
  This function is only useful if you accidentally close the GUI and want to reopen it, as it is normally launched upon library loading.
}
\value{
  None.
%%  If it is a LIST, use
%%  \item{}{}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\author{
Yvonnick Noel, University of Brittany, Rennes, France.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ GUI }

