% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_wtd_summarise.R
\name{full_wtd_summarise}
\alias{full_wtd_summarise}
\title{Get weighted mean and quantiles by group}
\usage{
full_wtd_summarise(data, vars, groups, weights)
}
\arguments{
\item{data}{A dataframe}

\item{vars}{A character vector of variables to summarise}

\item{groups}{A character vector of variables to group by}

\item{weights}{The weights variable}
}
\value{
A dataframe
}
\description{
Get weighted mean and quantiles by group
}
\examples{
full_wtd_summarise(data = penguins, vars = "bill_len", groups = "species", weights = "body_mass")
}
