% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pcs_coord.R
\name{get_pcs_coord}
\alias{get_pcs_coord}
\title{Get PCS coord after dimension reduction}
\usage{
get_pcs_coord(data, var)
}
\arguments{
\item{data}{a dataframe with dimensions and pcs grouping variables}

\item{var}{a pcs niv}
}
\value{
a summarised dataframe
}
\description{
Get PCS coord after dimension reduction
}
