% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_x_keuros.R
\name{label_x_keuros}
\alias{label_x_keuros}
\title{Title Label x axis in k euros}
\usage{
label_x_keuros()
}
\value{
a ggplot object
}
\description{
Title Label x axis in k euros
}
\details{
Some graphic devices may require the \pkg{showtext} package
to correctly display the euro symbol.
}
\examples{
if (require("showtext")) {
library(ggplot2)
showtext_auto()
ggplot(penguins, aes(x = bill_len, y = flipper_len)) +
geom_point() +
label_y_keuros()
}
}
