% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_inside.R
\name{legend_inside}
\alias{legend_inside}
\title{Title}
\usage{
legend_inside(justification = c(0, 1))
}
\arguments{
\item{justification}{A vector of length 2}
}
\value{
A ggplot object with legend insid
}
\description{
Title
}
\details{
Cette fonction doit être utilisée AVANT theme_...()
Avec theme_set() toutefois, ça ne pose aucun problème.
}
\examples{
penguins |>
ggplot2::ggplot(ggplot2::aes(x = bill_len,
 y = flipper_len,
  colour = species)) +
  legend_inside()
}
