% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Penguins dataset by Allison Horst}
\format{
A data frame with X rows and Y variables:
}
\source{
palmerpenguins package
}
\usage{
penguins
}
\description{
Penguins dataset by Allison Horst
}
\keyword{datasets}
