% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_acm_data.R
\name{prep4acm}
\alias{prep4acm}
\title{Préparer des données pour une ACM}
\usage{
prep4acm(data)
}
\arguments{
\item{data}{data.frame contenant uniquement les variables
actives de l'ACM, prélablement sélectionnées.}
}
\value{
data.frame prêt pour FactoMineR::MCA
}
\description{
Préparer des données pour une ACM
}
