% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_acm_data.R
\name{prepare_acm_data}
\alias{prepare_acm_data}
\title{Préparer des données pour une ACM}
\usage{
prepare_acm_data(data, var_actives, var_sup = NULL)
}
\arguments{
\item{data}{data.frame source}

\item{var_actives}{variables actives}

\item{var_sup}{variables supplémentaires}
}
\value{
data.frame prêt pour FactoMineR::MCA
}
\description{
Préparer des données pour une ACM
}
