% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_cats.R
\name{rare_cats}
\alias{rare_cats}
\title{Détecter les modalités rares pour une ACM}
\usage{
rare_cats(data, min_pct = 5)
}
\arguments{
\item{data}{data.frame de variables qualitatives}

\item{min_pct}{seuil minimal de fréquence (\%)}
}
\value{
Un vecteur de modalités à exclure
}
\description{
Calcule les fréquences des modalités et retourne celles
à exclure (rares et/ou NA recodés).
}
