% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_acm.R
\name{run_acm}
\alias{run_acm}
\title{Exécuter une ACM}
\usage{
run_acm(data_prepare, excl = NULL, quali_sup = NULL, ncp = 6)
}
\arguments{
\item{data_prepare}{données préparées}

\item{excl}{fonction ou vecteur de modalités à exclure}

\item{quali_sup}{indices des variables supplémentaires}

\item{ncp}{nombre de dimensions}
}
\value{
objet FactoMineR::MCA
}
\description{
Exécuter une ACM
}
