% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_acm.R
\name{shape_acm}
\alias{shape_acm}
\title{Pipeline complet ACM}
\usage{
shape_acm(
  data,
  var_actives,
  var_sup = NULL,
  excl = atelier::rare_cats,
  ncp = 6
)
}
\arguments{
\item{data}{data.frame source}

\item{var_actives}{variables actives}

\item{var_sup}{variables supplémentaires}

\item{excl}{fonction ou vecteur de modalités exclues}

\item{ncp}{nombre de dimensions}
}
\value{
objet FactoMineR::MCA
}
\description{
Pipeline complet ACM
}
