% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblk_summary.R
\name{tblk_summary}
\alias{tblk_summary}
\title{Personnalised tbl_summary}
\usage{
tblk_summary(data, vars, by = NULL, missing_text = "Manquant")
}
\arguments{
\item{data}{A dataframe}

\item{vars}{A named list, with variable names as names,
and labels}

\item{by}{Optionnal grouping variable}

\item{missing_text}{Label for missing text}
}
\value{
A gtsummary::tbl_summary()
}
\description{
Personnalised tbl_summary
}
