% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_acm.R
\name{tidy_acm}
\alias{tidy_acm}
\title{Construire un tableau consolidé des résultats d'ACM}
\usage{
tidy_acm(res_acm)
}
\arguments{
\item{res_acm}{Objet résultat d'une ACM (ex: FactoMineR::MCA)}
}
\value{
Un data.frame contenant les modalités actives,
exclues et supplémentaires.
}
\description{
Construire un tableau consolidé des résultats d'ACM
}
