% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_vars.R
\name{tidy_vars}
\alias{tidy_vars}
\title{tidy active vars MCA metrics}
\usage{
tidy_vars(x)
}
\arguments{
\item{x}{Le résultat de FactoMineR::MCA}
}
\value{
Un data.frame avec toutes les informations sur les
variables/modalités actives
}
\description{
tidy active vars MCA metrics
}
