% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm.R
\name{acm}
\alias{acm}
\title{Title}
\usage{
acm(data, min_pct)
}
\arguments{
\item{data}{A dataframe with active variables only}

\item{min_pct}{Minimum threshold to keep modalities}
}
\value{
A FactoMineR MCA
}
\description{
Title
}
