\name{AtelieR-package}
\alias{AtelieR-package}
\docType{package}
\title{
A GTK GUI for a set of sampling and bayesian inference tools}
\description{
AtelieR is a GTK interface for a bunch of statistical workshops. Suited for teaching purposes. It currently contains the following modules:\cr
- construction of the normal distribution (Central Limit Theorem)
- effect of a change of variable on the shape of gaussian distributions,
- distribution of a sample mean,
- effect of df correction on the expectation of the sample variance distribution
- distribution of a sample variance
- a probability calculator from various distributions,
- bayesian test on a proportion, credibility interval, and posterior probability of an hypothesis,
- bayesian test on a mean, credibility interval, and posterior probability of an hypothesis,
- bayesian test on a variance, credibility interval, and posterior probability of an hypothesis,
- bayesian test for comparing several proportions,
- bayesian test for the analysis of contingency tables,
- bayesian test for comparing several means.}
\details{
\tabular{ll}{
Package: \tab AtelieR\cr
Type: \tab Package\cr
Version: \tab 0.22\cr
Date: \tab 2011-11-10\cr
License: \tab GPL 2.0\cr
}
Just type AtelieR() at the prompt to launch the GUI.
}
\author{
Maintainer: Yvonnick Noel <yvonnick.noel@univ-rennes2.fr>
}

