% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_excl.R
\name{tidy_excl_coord}
\alias{tidy_excl_coord}
\title{tidy excluded modalities}
\usage{
tidy_excl_coord(x)
}
\arguments{
\item{x}{A MCA object}
}
\value{
a dataframe of excl coordinates
}
\description{
tidy excluded modalities
}
